+++ rust/hyperlane-core/src/chain.rs
@@ -80,6 +80,8 @@ pub enum KnownHyperlaneDomain {
 
     Neutron = 1853125230,
 
+    Osmosis = 875,
+
     Injective = 6909546,
     InEvm = 2525,
 
@@ -96,6 +98,12 @@ pub enum KnownHyperlaneDomain {
     Zetachain = 7000,
 
     PlumeTestnet = 161221135,
+
+    Fraxtal = 252,
+
+    Linea = 59144,
+
+    Sei = 1329,
 
     // -- Local test chains --
     /// Test1 local chain
@@ -229,7 +237,7 @@ impl KnownHyperlaneDomain {
             Mainnet: [
                 Ethereum, Avalanche, Arbitrum, Polygon, Optimism, BinanceSmartChain, Celo,
                 Moonbeam, Gnosis, MantaPacific, Neutron, Injective, InEvm, Ancient8, Blast,
-                Mode, Redstone, Viction, Zetachain
+                Mode, Redstone, Viction, Zetachain, Fraxtal, Linea, Sei, Osmosis
             ],
             Testnet: [
                 Alfajores, MoonbaseAlpha, Sepolia, ScrollSepolia, Chiado, PlumeTestnet, Fuji, BinanceSmartChainTestnet, Holesky
@@ -247,11 +255,12 @@ impl KnownHyperlaneDomain {
                 Optimism, BinanceSmartChain, BinanceSmartChainTestnet, Celo, Gnosis,
                 Alfajores, Moonbeam, InEvm, Ancient8, Blast, Mode, Redstone, Viction,
                 Zetachain, MoonbaseAlpha, ScrollSepolia, Chiado, MantaPacific, PlumeTestnet,
+                Fraxtal, Linea, Sei,
                 Test1, Test2, Test3
             ],
             HyperlaneDomainProtocol::Fuel: [FuelTest1],
             HyperlaneDomainProtocol::Sealevel: [SealevelTest1, SealevelTest2],
-            HyperlaneDomainProtocol::Cosmos: [CosmosTest99990, CosmosTest99991, Neutron, Injective],
+            HyperlaneDomainProtocol::Cosmos: [CosmosTest99990, CosmosTest99991, Neutron, Osmosis, Injective],
         })
     }
 
@@ -264,7 +273,7 @@ impl KnownHyperlaneDomain {
                 Ethereum, Sepolia, Holesky, Polygon, Avalanche, Fuji, Optimism,
                 BinanceSmartChain, BinanceSmartChainTestnet, Celo, Gnosis, Alfajores, Moonbeam, MoonbaseAlpha,
                 ScrollSepolia, Chiado, MantaPacific, Neutron, Injective, InEvm, Ancient8, Blast, Mode, Redstone,
-                Viction, Zetachain,
+                Viction, Zetachain, Fraxtal, Linea, Sei, Osmosis,
                 Test1, Test2, Test3, FuelTest1, SealevelTest1, SealevelTest2, CosmosTest99990, CosmosTest99991
             ],
         })